+++ changelog.md
@@ -13,6 +13,7 @@
   that template instead of the default database.
 * Made it possible to drag and drop rows in the desired order.
 * Fixed bug where the rows could get out of sync during real time collaboration.
+* Made it possible to export and import the file field including contents.
 * Added `fill_users` admin management command which fills baserow with fake users.
 * **Premium**: Added user admin area allowing management of all baserow users.